# 4310 Do not count order shipments in Order#shipping_method